+++ packages/terra-core-docs/src/terra-dev-site/doc/dropdown-button/example/DefaultSplitButton.jsx
@@ -9,6 +9,9 @@ const Example = () => {
       <SplitButton
         primaryOptionLabel="Reply"
         onSelect={() => setMessage('Reply clicked')}
+        inputAttrs={{
+          'aria-label': 'default',
+        }}
       >
         <Item label="Reply All" onSelect={() => setMessage('Reply All clicked')} />
         <Item label="Forward" onSelect={() => setMessage('Forward clicked')} />